Übersetzter Text Test added.
The TimeSlots class has been adapted and please note that I have added a loop in the constructor: the loop runs through each element of the timeSlots array and sets it to TimeSlotValue.FREE to ensure that all time slots are free at the beginning.

BasicStudyplanAlgorithm still needs to be tested, I think it makes more sense if @jilzerndt tests it, since she implemented it.

Interfaces do not need to be tested, so I have deleted the prepared interface tests.